"Fix" 커리어 프로필 작성 사이드바 분리, 수정
